moved ResurfaceWarning class to a new file
